remove "Plan Chess From Outer Space" subdirectory
This commit moves the project root directory to the repo root.

The project root formerly lies in a subdirectory of the repo with
spaces; this unnecessarily complicates navigation through web
interfaces and the command line, and is also not in line with
common practice.

Moving everything out of the subdirectory aligns this repository with
the practice of most significant libre software repositories, and also
potentially eases navigation.