add GetMoves interface to BoardManager
This commit adds the GetMoves function to BoardManager. It takes a
position as input and returns an IEnumerable of positions it can
move to.

The interface defers the actual task of enumerating positions to the
ChessPiece instance found at said position, as implementing piece move
in BoardManager would be highly detrimental to code quality :)