Make ChessPiece abstract
There are so many kinds of movesets to implement; using a large switch
block is an option, but would make our code difficult to maintain.
Instead I am opting to subclass ChessPiece; for each moveset there
will be a ChessPiece subclass, with the moveset-specific functions
implemented.

ChessPiece.GetMoves is turned into an abstract function.