lower capture check to ChessPiece
The coordinates based abstractions are largely neglected in this project
in favor of directly manipulating ChessPieces and BoardSquares, so it
don't really make sense to only check captures in BoardManager's move
function.

This patch moves the capture check down to ChessPiece. Additionally,
the capture check no longer immediately calls Destroy() on the captured
piece as other objects can still hold references to it for an indeter-
minate amount of time, and a dead flag is set instead.