amend! Implement basic AI class
Implement basic AI class

The AI class is the bot that plays all the piece moves. It has two
functions: SyncAI for resetting the AI and restart calculation, and
RecommendMove for carrying out the actual move.

Currently the move calculations are pretty simple and are done
synchronously. The interface design is in place to allow a minimax
alpha-beta search to drop in in the future.

----

Amends:
 1. turn AI into a service.
 2. Fix score ordering. SortedSet<T> orders smallest element (as
    determined via IComparable<T>.CompareTo()) first, while we want
    to pick the highest scored move. This makes the AI behavior work
    as intended, though piece decisions can feel overly aggresive.