timeline: fix race condition on killing a piece
A TimelineNode's kill animation deletes the node itself after the
animation concludes. However, the UnitOrderTimelineController keeps
a reference to that around while waiting for it to conclude, creating
a race condition that can result in crash.

The race condition is fixed by splicing the TimelineNode reference
from UnitOrderTimelineController's piece list before the animation and
the Unity coroutine wait. This way, all other references are actually
dropped before the TimelineNode deletes itself, and the random crash
will no longer happen.